+++ training_mod_tui/src/lib.rs
@@ -4,198 +4,16 @@ use tui::{
     layout::{Constraint, Corner, Direction, Layout},
     style::{Color, Modifier, Style},
     text::Spans,
-    widgets::{Block, List, ListItem, ListState},
+    widgets::{Tabs, Paragraph, Block, List, ListItem, ListState},
     Frame,
 };
 
 pub use tui::{backend::TestBackend, Terminal};
-use tui::widgets::Tabs;
 use std::collections::HashMap;
 
-pub struct MultiStatefulList<T> {
-    pub lists: Vec<StatefulList<T>>,
-    pub state: usize,
-    pub total_len: usize
-}
+mod list;
 
-impl<T: Clone> MultiStatefulList<T> {
-    pub fn selected_list_item(&mut self) -> &mut T {
-        let (list_section, list_idx) = self.idx_to_list_idx(self.state);
-        &mut self
-            .lists[list_section]
-            .items[list_idx]
-    }
-
-    fn idx_to_list_idx(&self, idx: usize) -> (usize, usize) {
-        for list_section in 0..self.lists.len() {
-            let list_section_min_idx = (self.total_len / self.lists.len()) * list_section;
-            let list_section_max_idx = (self.total_len / self.lists.len()) * (list_section + 1);
-            if (list_section_min_idx..list_section_max_idx).contains(&idx) {
-                return (list_section, idx - list_section_min_idx)
-            }
-        }
-        (0, 0)
-    }
-
-    fn list_idx_to_idx(&self, list_idx: (usize, usize)) -> usize {
-        let list_section = list_idx.0;
-        let mut list_idx = list_idx.1;
-        for list_section in 0..list_section {
-            list_idx += self.lists[list_section].items.len();
-        }
-        list_idx
-    }
-
-    pub fn with_items(items: Vec<T>, num_lists: usize) -> MultiStatefulList<T> {
-        let lists = (0..num_lists).map(|list_section| {
-            let list_section_min_idx = (items.len() / num_lists) * list_section;
-            let list_section_max_idx = (items.len() / num_lists) * (list_section + 1);
-            let mut state = ListState::default();
-            if list_section == 0 {
-                // Enforce state as first of list
-                state.select(Some(0));
-            }
-            StatefulList {
-                state: state,
-                items: items[list_section_min_idx..list_section_max_idx].to_vec(),
-            }
-        }).collect();
-        let total_len = items.len();
-        MultiStatefulList {
-            lists: lists,
-            total_len: total_len,
-            state: 0
-        }
-    }
-
-    pub fn next(&mut self) {
-        let (list_section, _) = self.idx_to_list_idx(self.state);
-        let (next_list_section, next_list_idx) = self.idx_to_list_idx(self.state+1);
-
-        if list_section != next_list_section {
-            self.lists[list_section].unselect();
-        }
-        let state;
-        if self.state + 1 >= self.total_len {
-            state = (0, 0);
-        } else {
-            state = (next_list_section, next_list_idx);
-        }
-
-        self.lists[state.0].state.select(Some(state.1));
-        self.state = self.list_idx_to_idx(state);
-    }
-
-    pub fn previous(&mut self) {
-        let (list_section, _) = self.idx_to_list_idx(self.state);
-        let (last_list_section, last_list_idx) = (self.lists.len() - 1, self.lists[self.lists.len() - 1].items.len() - 1);
-
-        self.lists[list_section].unselect();
-        let state;
-        if self.state == 0 {
-            state = (last_list_section, last_list_idx);
-        } else {
-            let (prev_list_section, prev_list_idx) = self.idx_to_list_idx(self.state - 1);
-            state = (prev_list_section, prev_list_idx);
-        }
-
-        self.lists[state.0].state.select(Some(state.1));
-        self.state = self.list_idx_to_idx(state);
-    }
-
-    pub fn next_list(&mut self) {
-        let (list_section, list_idx) = self.idx_to_list_idx(self.state);
-        let next_list_section = (list_section + 1) % self.lists.len();
-        let next_list_idx;
-        if list_idx > self.lists[next_list_section].items.len() - 1 {
-            next_list_idx = self.lists[next_list_section].items.len() - 1;
-        } else {
-            next_list_idx = list_idx;
-        }
-
-        if list_section != next_list_section {
-            self.lists[list_section].unselect();
-        }
-        let state = (next_list_section, next_list_idx);
-
-        self.lists[state.0].state.select(Some(state.1));
-        self.state = self.list_idx_to_idx(state);
-    }
-
-    pub fn previous_list(&mut self) {
-        let (list_section, list_idx) = self.idx_to_list_idx(self.state);
-        let prev_list_section;
-        if list_section == 0 {
-            prev_list_section = self.lists.len() - 1;
-        } else {
-            prev_list_section = list_section - 1;
-        }
-
-        let prev_list_idx;
-        if list_idx > self.lists[prev_list_section].items.len() - 1 {
-            prev_list_idx = self.lists[prev_list_section].items.len() - 1;
-        } else {
-            prev_list_idx = list_idx;
-        }
-
-        if list_section != prev_list_section {
-            self.lists[list_section].unselect();
-        }
-        let state = (prev_list_section, prev_list_idx);
-
-        self.lists[state.0].state.select(Some(state.1));
-        self.state = self.list_idx_to_idx(state);
-    }
-}
-
-pub struct StatefulList<T> {
-    pub state: ListState,
-    pub items: Vec<T>,
-}
-
-impl<T> StatefulList<T> {
-    pub fn with_items(items: Vec<T>) -> StatefulList<T> {
-        let mut state = ListState::default();
-        // Enforce state as first of list
-        state.select(Some(0));
-        StatefulList {
-            state: state,
-            items,
-        }
-    }
-
-    pub fn next(&mut self) {
-        let i = match self.state.selected() {
-            Some(i) => {
-                if i >= self.items.len() - 1 {
-                    0
-                } else {
-                    i + 1
-                }
-            }
-            None => 0,
-        };
-        self.state.select(Some(i));
-    }
-
-    pub fn previous(&mut self) {
-        let i = match self.state.selected() {
-            Some(i) => {
-                if i == 0 {
-                    self.items.len() - 1
-                } else {
-                    i - 1
-                }
-            }
-            None => 0,
-        };
-        self.state.select(Some(i));
-    }
-
-    pub fn unselect(&mut self) {
-        self.state.select(None);
-    }
-}
+use crate::list::{StatefulList, MultiStatefulList};
 
 /// We should hold a list of SubMenus.
 /// The currently selected SubMenu should also have an associated list with necessary information.
@@ -258,12 +76,10 @@ impl<'a> App<'a> {
         tabs.insert("Defensive Settings", vec![]);
         tabs.insert("Miscellaneous Settings", vec![]);
 
-        let mut menu_items = Vec::new();
         for sub_menu in menu.sub_menus.iter() {
             for tab_spec in tab_specifiers.iter() {
                 if tab_spec.1.contains(&sub_menu.title) {
                     tabs.get_mut(tab_spec.0).unwrap().push(sub_menu.clone());
-                    menu_items.push( sub_menu.clone());
                 }
             }
         };
@@ -276,10 +92,8 @@ impl<'a> App<'a> {
                 MultiStatefulList::with_items(tabs.get(k).unwrap().clone(), num_lists)
             );
         });
-        let mut tabs_sorted = tabs.keys().cloned().collect::<Vec<&str>>();
-        tabs_sorted.sort();
         let mut app = App {
-            tabs: StatefulList::with_items(tabs_sorted),
+            tabs: StatefulList::with_items(tab_specifiers.iter().map(|(tab_title, _)| *tab_title).collect()),
             menu_items: menu_items_stateful,
             selected_sub_menu_toggles: MultiStatefulList::with_items(vec![], 0),
             selected_sub_menu_onoff_selectors: MultiStatefulList::with_items(vec![], 0),
@@ -327,63 +141,39 @@ impl<'a> App<'a> {
 
     pub fn sub_menu_next(&mut self) {
         match SubMenuType::from_str(self.sub_menu_selected()._type) {
-            SubMenuType::TOGGLE => {
-                self.selected_sub_menu_toggles.next();
-            },
-            SubMenuType::SLIDER => {
-                self.selected_sub_menu_sliders.next();
-            },
-            SubMenuType::ONOFF => {
-                self.selected_sub_menu_onoff_selectors.next()
-            },
+            SubMenuType::TOGGLE => self.selected_sub_menu_toggles.next(),
+            SubMenuType::SLIDER => self.selected_sub_menu_sliders.next(),
+            SubMenuType::ONOFF => self.selected_sub_menu_onoff_selectors.next(),
         }
     }
 
     pub fn sub_menu_next_list(&mut self) {
         match SubMenuType::from_str(self.sub_menu_selected()._type) {
-            SubMenuType::TOGGLE => {
-                self.selected_sub_menu_toggles.next_list();
-            },
-            SubMenuType::SLIDER => {
-                self.selected_sub_menu_sliders.next_list();
-            },
-            SubMenuType::ONOFF => {
-                self.selected_sub_menu_onoff_selectors.next_list()
-            },
+            SubMenuType::TOGGLE => self.selected_sub_menu_toggles.next_list(),
+            SubMenuType::SLIDER => self.selected_sub_menu_sliders.next_list(),
+            SubMenuType::ONOFF => self.selected_sub_menu_onoff_selectors.next_list(),
         }
     }
 
     pub fn sub_menu_previous(&mut self) {
         match SubMenuType::from_str(self.sub_menu_selected()._type) {
-            SubMenuType::TOGGLE => {
-                self.selected_sub_menu_toggles.previous();
-            },
-            SubMenuType::SLIDER => {
-                self.selected_sub_menu_sliders.previous();
-            },
-            SubMenuType::ONOFF => {
-                self.selected_sub_menu_onoff_selectors.previous()
-            },
+            SubMenuType::TOGGLE => self.selected_sub_menu_toggles.previous(),
+            SubMenuType::SLIDER => self.selected_sub_menu_sliders.previous(),
+            SubMenuType::ONOFF => self.selected_sub_menu_onoff_selectors.previous(),
         }
     }
 
     pub fn sub_menu_previous_list(&mut self) {
         match SubMenuType::from_str(self.sub_menu_selected()._type) {
-            SubMenuType::TOGGLE => {
-                self.selected_sub_menu_toggles.previous_list();
-            },
-            SubMenuType::SLIDER => {
-                self.selected_sub_menu_sliders.previous_list();
-            },
-            SubMenuType::ONOFF => {
-                self.selected_sub_menu_onoff_selectors.previous_list()
-            },
+            SubMenuType::TOGGLE => self.selected_sub_menu_toggles.previous_list(),
+            SubMenuType::SLIDER => self.selected_sub_menu_sliders.previous_list(),
+            SubMenuType::ONOFF => self.selected_sub_menu_onoff_selectors.previous_list(),
         }
     }
 
-    pub fn sub_menu_strs_and_states(&mut self) -> (&str, Vec<(Vec<(&str, &str)>, ListState)>) {
-        (self.sub_menu_selected().title,
-        match SubMenuType::from_str(self.sub_menu_selected()._type) {
+    pub fn sub_menu_strs_and_states(&mut self) -> (&str, &str, Vec<(Vec<(&str, &str)>, ListState)>) {
+        (self.sub_menu_selected().title, self.sub_menu_selected().help_text,
+         match SubMenuType::from_str(self.sub_menu_selected()._type) {
             SubMenuType::TOGGLE => {
                 self.selected_sub_menu_toggles.lists.iter().map(|toggle_list| {
                     (toggle_list.items.iter().map(
@@ -392,9 +182,6 @@ impl<'a> App<'a> {
                 }).collect()
             },
             SubMenuType::SLIDER => {
-                // self.selected_sub_menu_sliders.items.iter().map(
-                //     |_slider| ("TODO", "TODO")
-                // ).collect()
                 vec![(vec![], ListState::default())]
             },
             SubMenuType::ONOFF => {
@@ -405,26 +192,6 @@ impl<'a> App<'a> {
                 }).collect()
             },
         })
-    }
-
-    pub fn sub_menu_state(&mut self) -> &ListState {
-        match SubMenuType::from_str(self.sub_menu_selected()._type) {
-            SubMenuType::TOGGLE => {
-                let (list_section, _) =
-                    self.selected_sub_menu_toggles.idx_to_list_idx(self.selected_sub_menu_toggles.state);
-                &self.selected_sub_menu_toggles.lists[list_section].state
-            },
-            SubMenuType::SLIDER => {
-                let (list_section, _) =
-                    self.selected_sub_menu_sliders.idx_to_list_idx(self.selected_sub_menu_sliders.state);
-                &self.selected_sub_menu_sliders.lists[list_section].state
-            },
-            SubMenuType::ONOFF => {
-                let (list_section, _) =
-                    self.selected_sub_menu_onoff_selectors.idx_to_list_idx(self.selected_sub_menu_onoff_selectors.state);
-                &self.selected_sub_menu_onoff_selectors.lists[list_section].state
-            },
-        }
     }
 
     pub fn on_a(&mut self) {
@@ -526,35 +293,48 @@ impl<'a> App<'a> {
 
 pub fn ui<B: Backend>(f: &mut Frame<B>, app: &mut App) -> String {
     let app_tabs = &app.tabs;
-    let titles = app_tabs.items.iter().cloned().map(Spans::from).collect();
+    let tab_selected = app_tabs.state.selected().unwrap();
+    let titles = app_tabs.items.iter().cloned().enumerate().map(|(idx, tab)|{
+        if idx == tab_selected {
+            Spans::from(">> ".to_owned() + tab)
+        } else {
+            Spans::from("  ".to_owned() + tab)
+        }
+    }).collect();
     let tabs = Tabs::new(titles)
         .block(Block::default().title("Ultimate Training Modpack Menu"))
         .style(Style::default().fg(Color::White))
         .highlight_style(Style::default().fg(Color::Yellow))
         .divider("|")
-        .select(app_tabs.state.selected().unwrap());
+        .select(tab_selected);
 
-    let tab_chunks = Layout::default()
+    let vertical_chunks = Layout::default()
         .direction(Direction::Vertical)
-        .constraints([Constraint::Percentage(30), Constraint::Percentage(70)].as_ref())
+        .constraints([Constraint::Percentage(30), Constraint::Percentage(50), Constraint::Percentage(20)].as_ref())
         .split(f.size());
 
     let list_chunks = Layout::default()
         .direction(Direction::Horizontal)
         .constraints([Constraint::Percentage(33), Constraint::Percentage(32), Constraint::Percentage(33)].as_ref())
-        .split(tab_chunks[1]);
+        .split(vertical_chunks[1]);
 
-    f.render_widget(tabs, tab_chunks[0]);
+    f.render_widget(tabs, vertical_chunks[0]);
 
     if app.outer_list {
         let tab_selected = app.tab_selected();
+        let mut item_help = None;
         for list_section in 0..app.menu_items.get(tab_selected).unwrap().lists.len() {
             let stateful_list = &app.menu_items.get(tab_selected).unwrap().lists[list_section];
             let items: Vec<ListItem> = stateful_list
                 .items
                 .iter()
                 .map(|i| {
-                    let lines = vec![Spans::from(i.title)];
+                    let lines = vec![Spans::from(
+                        if stateful_list.state.selected().is_some() {
+                            i.title.to_owned()
+                        } else {
+                            "   ".to_owned() + i.title
+                        })];
                     ListItem::new(lines).style(Style::default().fg(Color::Black).bg(Color::White))
                 })
                 .collect();
@@ -569,11 +349,20 @@ pub fn ui<B: Backend>(f: &mut Frame<B>, app: &mut App) -> String {
                 .highlight_symbol(">> ");
 
             let mut state = stateful_list.state.clone();
+            if state.selected().is_some() {
+                item_help = Some(stateful_list.items[state.selected().unwrap()].help_text);
+            }
 
             f.render_stateful_widget(list, list_chunks[list_section], &mut state);
         }
+
+        let help_paragraph = Paragraph::new(
+            item_help.unwrap_or("").replace("\"", "") +
+            "\nA: Enter sub-menu | B: Exit menu | ZL/ZR: Next tab | R: Save defaults"
+        );
+        f.render_widget(help_paragraph, vertical_chunks[2]);
     } else {
-        let (title, mut sub_menu_str_lists) = app.sub_menu_strs_and_states();
+        let (title, help_text, mut sub_menu_str_lists) = app.sub_menu_strs_and_states();
         for list_section in 0..sub_menu_str_lists.len() {
             let sub_menu_str = sub_menu_str_lists[list_section].0.clone();
             let sub_menu_state = &mut sub_menu_str_lists[list_section].1;
@@ -596,6 +385,12 @@ pub fn ui<B: Backend>(f: &mut Frame<B>, app: &mut App) -> String {
                 .highlight_symbol(">> ");
             f.render_stateful_widget(values_list, list_chunks[list_section], sub_menu_state);
         }
+
+        let help_paragraph = Paragraph::new(
+            help_text.replace("\"", "") +
+                "\nA: Select toggle | B: Exit submenu"
+        );
+        f.render_widget(help_paragraph, vertical_chunks[2]);
     }
